Forecasting_Assignment_2
Welcome to assignment 2. The objective of task 1 of this assignment is to analyze and forecast the amount of horizontal solar radiation reaching the ground at a particular location over the globe.
For this aim,we need to work on the monthly average horizontal solar radiation and the monthly precipitation series measured at the same location between January 1960 and December 2014.
To carry out this investigation I will be using R which is an open course statistical programming tool to examine and inspect the various elements of the time series
2
regression methods in which I will be using (dLagM package), dynamic linear models (dynlm package), and exponential smoothing and corresponding state-space models.