Time Series_Assignment_1
The ozone layer is what protects us living being from harmful UV rays . Ultraviolet (UV) radiation is a type of energy produced by the sun and some artificial sources, such as solariums. The sun’s ultraviolet (UV) radiation is the main cause of skin cancer. Thus if the ozone layer delpletes it is of great concern to all of us. The dataset being used will be used to analyse yearly changes in the thickness of Ozone layer 1927 to 2016 in Dobson units. A negative value in the dataset represents a decrease in the thickness and a positive value represents an increase in the thickness.

We need to predict the yearly changes for the next 5 years (2017-2022) using the best fitting model from among the linear, quadratic, cosine, cyclical or seasonal trend model , with the help of the model building strategy which includes descriptive look, model specification (or identification), model fitting, model diagnostics, and last but not the least forecasting, which is the main goal of the task