Time Series_Assignment_2
In this assignment task we need to analyze the eggs depositions (in millions) of age-3 Lake Huron Bloaters (Coregonus hoyi) data set between the years 1981 and 1996 , using the analysis methods covered in the course and applying suitable approaches to determine the best fitting model from a set of possible models to forecast the egg depositions for the next 5 years.

Our first task will be to covert the data from a comma separate file to a time series object in R . 
Once the time series has been created for the data, we conduct some preliminary visual inspection to check the distribution of the data and the skewness . This gives us a general idea about the general variance in the data set and how the data is dispersed . 
We then analyze the time series for trend , seasonality , intervention points and whether the series is Autoregressive or has a moving average component along with the scatter plot (lag plot ) to check whether there is any correlation between the data and a its copy one time period apart. 
ACF and PACF plots yield symptoms of non-stationarity which can be dealt with through differencing  but if  there is changing variance in the series i.e. difference in values between successive points is not constant then Box-Cox or LOG transformations can be carried out before the differencing to stabilize the series . 
Augmented Dickey-Fuller Test checks whether differencing results in a stationary behavior by yielding a p-value . IF the p-value is less than 0.05 we reject the null hypotheses that the series is non-stationary. Differencing of the series is carried till we arrive at a p-value that is less than 0.05
ACF and PACF plots of the differenced series are analyzed to determine the order of the ARIMA series