Time Series_Final_Project
## 1. Introduction


#### 1.1 Problem Statement

The aim of this report is to explore air quality in Beijing as well as propose a suitable fitting time series model for the dataset, through the implementation of several refined techniques. To achieve this a traditional model building strategy shall be implemented with the following 3 steps; model specification, model fitting and model diagnostic. In the model specification step the appropriate orders shall be determined for each of the auto-regressive (AR), moving average (MA) and difference components. In the model fitting step parameter estimation shall be performed as to fit each coefficient of the generic models determined in the specification step. Finally, in the model diagnostic step the validity of the best performing models will be assessed. The selected model will be the strongest model that additionally satisfies the residual analysis.
Once the strongest model has been determined from the candidate set, it shall be used for providing a forecast of the air quality in Beijing over the next 10 months.


#### 1.2 Background

A typical method for determining a nation’s air quality ranking is through measuring the concentration of a common air pollutant known as PM2.5. PM2.5 are fine particles which have a diameter measuring 2.5µm or less. Particles of this size are typically chosen as indicators of air quality due to particle size and respiratory risk being inversely correlated (EPA Victoria, 2019). PM2.5 are particularly notorious for readily infiltrating the lungs and in rare cases even passing through the bloodstream (EPA Victoria, 2019).
The population of China has long been notorious for enduring one of the world’s worst air quality. Extensive quantities of air pollutants have long found china possessing one of the lowest Environmental Sustainability Index across the globe. As recently as 2015, 16 of the top 20 cities with the highest levels of air pollution have been located in china (Ye. M, 2015). The surge in China’s widespread pollution can be attributed to numerous factors including: a significant increase in population, manufacturing, industrialisation and traffic congestion.
The capital Beijing has found itself in this list and has long held the reputation of one of the world’s most polluted cities. However as of 2020 this reputation seems to be somewhat undeserved with significant improvements in the air quality having been recorded over the last few years. There have not only been improvements but these improvements have proved significant with the average annual PM2.5 more than halving since 2009 (IQAIR, 2020). These improvements have been made possible through heavy regulation and concerted efforts by Beijing’s policymakers in the pursuit of meeting the World Health Organisation (WHO) guideline of a PM2.5 concentration level of 10µg/m3 (IQAIR, 2020). These regulations have included methods such as significantly limiting the issue of licenses down to only 100,000 applicants per year, halting deforestation as well as transitioning away from coal in favour of natural gas (IQAIR, 2020). The measures have certainly proven effective with Beijing recording an all-time average low of 23 µg/m3 in the month of August 2019 and since 2018 managing to remain outside the top 100 most polluted cities (IQAIR, 2020). However, despite all the success Beijing has experienced, it still typically averages more than 4 times that of the WHO recommended level.


#### 1.3 Data Description


The dataset used in this report records the daily levels of PM2.5 measured in µg/m3 that were present in the Chinese capital Beijing from the 01/01/2014 through to the 21/05/2020. This dataset was compiled through the use of source data recorded at the IQAir Beijing station. The IQAir Beijing station’s records are updated almost daily and can be obtained from the following link: https://www.iqair.com/china/beijing

The dataset proves to be a significant size containing exactly 2333 observations. Since the dataset is relatively large, significant confidence can be placed on the assumption of the data being normally distributed. This conclusion can be reached through the reliance on the Central Limit Theorem. The assumption of normality is essential in this report with most statistical methods being implemented such as the Bayesian Information Criterion (BIC) requiring this assumption to be valid. Since the normality assumption is of the utmost importance in statistical analysis extra care will still be taken even though the dataset is relatively large. To ensure the normality assumption is being upheld, Q-Q Normality plots and Shapiro-Wilks tests shall be explored throughout the report.

An additional boon associated with the size of the dataset is related to overfitting. Since the dataset is of a significant size the risk of overfitting is substantially reduced and hence more complicated models may be considered. However, as an additional measure to further minimise the risk of overfitting, a standard rule of at least 10 observations per model parameter shall be followed. This arbitrary rule shall only be violated if the model specification step provides significant support for contrary models. Additionally, following this rule will assist in ensuring that the Principal of Parsimony (PoP) is being sufficiently followed.

Finally, the compiled dataset requires some pre-processing with the dataset containing several missing values. Fortunately, the number of missing values is not too significant with most corresponding to major western public festivals such as Christmas and New Year.