Applied Research Project - Time_Series_Work_Integrated_Learning
William A McNeil & Co (WAM) is one of the largest modern and premium quality home decor & home wares supplier in Australia. This research paper intends to study the sales trend of WAM products from July 2017 to March 2020.
WAM sells product to its retail and wholesale customers as cash sale and credit sale. Company have multiple product line and various products grouped under each product line. Sales invoice of the company is the primary source of the data and it was used for analyzing trends.
For designing various stages of research cross-industry process for data mining (CRISP-DM) frame was followed. Data analysis was done in three stages. In the first stage exploratory analysis was done using bar graphs and Pareto analysis. Plots of quantities sold and sales, did not show any trend or pattern. Pareto analysis was used to identify the type of contributors and comparison of Pareto analysis was done to check if there was change in the type of popular products sold in each year. In the second stage, hypothesis testing was done for categorical variables and the results shows that there is no significant relationship between the variables. As the final step of data analysis, a time series analysis and forecasting model was done to identify the best models and to predict the sales value for next 10 months.
From the analysis it can be concluded that there is no consistent growth on the sales value. Considering the fact that there was no change in the internal environmental factors fluctuations on the sales can be related to external environmental factors. Pareto analysis shows that there is not much change in the popular products in various years. Among the 55 products being sold 15 products contribute to 80% of sales value and hence focusing on these 15 products will help to optimize the production. Forecast values for the next 10 months are positive. One of the major limitations of the study is that reason affecting the trend was not considered.